requirements.txt
Flask==2.2.2
Flask_Caching==2.0.1
Flask_Login==0.6.2
flask_sqlalchemy==3.0.3
Pillow==9.4.0
Werkzeug==2.2.2
